@@ -9,12 +9,11 @@ import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.AsyncTask;
 import android.os.Bundle;
-import android.support.v4.widget.DrawerLayout;
+import androidx.drawerlayout.widget.DrawerLayout;
 import android.util.Log;
 import android.view.View;
 import android.view.WindowManager;
 import android.widget.Button;
-import android.widget.FrameLayout;
 import android.widget.ProgressBar;
 import android.widget.TextView;
 import android.widget.Toast;
@@ -24,7 +23,6 @@ import java.lang.reflect.InvocationTargetException;
 import benchmark.IBenchmark;
 import database.Database;
 import database.Score;
-import stopwatch.Timer;
 
 /**
  * Measures CPU performance.
